requirements.txt
customtkinter
psycopg2
python-dotenv
cryptography
pyperclip
pyinstaller